requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.4.26
charset-normalizer==3.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tzlocal==5.3.1
urllib3==2.4.0
